Create guide
This is a pure function because it does not rely on any external state or side effects - it simply generates a random pulse reading and returns it as an object with a timestamp. You can call this function multiple times to simulate fetching pulse data at different intervals.

Note that if you have an actual device for measuring pulse, you will need to use an appropriate library or API to communicate with the device and fetch data in real-time.